requirements.txt
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
blinker==1.7.0
boto3==1.34.144
botocore==1.34.144
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.3.1
huggingface-hub==0.20.3
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-generative-ai==2.1.1
ibm-watson-machine-learning==1.0.349
ibm_watsonx_ai==0.2.6
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kafka-python==2.0.2
kiwisolver==1.4.5
langchain==0.1.10
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.53
langchain-ibm==0.1.4
langchain-text-splitters==0.0.2
langchainhub==0.1.20
langsmith==0.1.3
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
openai==1.30.5
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.10.3
packaging==23.2
panda==0.3.1
pandas==1.5.3
paramiko==3.3.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.1.1
pypdf==4.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.2
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.3.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.26
stack-data==0.6.3
starlette==0.27.0
streamlit==1.31.1
streamlit-chat==0.1.1
streamlit-tags==1.2.8
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
text-generation==0.6.1
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.23.2
validators==0.22.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
